Change AudioSource::read to take AudioBuffer.
Format requirements will now be passed only from the sink to the
AudioSource implementations. The audio format is bundled with a
reference to a slice of samples in AudioBuffer.

+++ src/sdl_util.rs
@@ -5,7 +5,7 @@ use sdl2::audio::AudioSpec;
 impl From<AudioSpec> for AudioFormat {
     fn from(spec: AudioSpec) -> Self {
         AudioFormat {
-            channels: spec.channels as u32,
+            channels: spec.channels,
             sample_rate: spec.freq as u32,
         }
     }
